backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Google Gemini AI
google-generativeai==0.3.2

# Voice processing
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11

# Audio processing
numpy==1.24.3
scipy==1.11.3
librosa==0.10.1

# Async and web
asyncio==3.4.3
python-multipart==0.0.6
python-dotenv==1.0.0

# CORS and middleware
fastapi-cors==0.0.6

# WebSocket support
python-socketio==5.9.0

# JSON and data handling
pydantic==2.5.0

# Logging
loguru==0.7.2
